+++ CotEditor/Sources/DocumentViewController.swift
@@ -9,7 +9,7 @@
 //  ---------------------------------------------------------------------------
 //
 //  © 2004-2007 nakamuxu
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -85,8 +85,11 @@ final class DocumentViewController: NSSplitViewController, SyntaxParserDelegate,
         
         // observe defaults change
         self.defaultsObservers = [
-            UserDefaults.standard.observe(key: .theme) { [unowned self] _ in
-                self.setTheme(name: ThemeManager.shared.defaultSettingName())
+            UserDefaults.standard.observe(key: .theme, options: [.old, .new]) { [unowned self] change in
+                guard change.old == nil || self.theme?.name == change.old else { return }
+                
+                let themeName = ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
+                self.setTheme(name: themeName)
             },
             UserDefaults.standard.observe(key: .showInvisibles, options: [.new]) { [unowned self] change in
                 self.showsInvisibles = change.new!
@@ -120,8 +123,6 @@ final class DocumentViewController: NSSplitViewController, SyntaxParserDelegate,
     override class var restorableStateKeyPaths: [String] {
         
         return super.restorableStateKeyPaths + [
-            #keyPath(isStatusBarShown),
-            #keyPath(showsNavigationBar),
             #keyPath(showsLineNumber),
             #keyPath(showsPageGuide),
             #keyPath(showsInvisibles),
@@ -437,7 +438,7 @@ final class DocumentViewController: NSSplitViewController, SyntaxParserDelegate,
     
     
     /// Whether status bar is visible
-    @objc var isStatusBarShown: Bool {
+    var isStatusBarShown: Bool {
         
         get {
             return self.statusBarItem?.isCollapsed == false
@@ -451,7 +452,7 @@ final class DocumentViewController: NSSplitViewController, SyntaxParserDelegate,
     
     
     /// visibility of navigation bars
-    @objc var showsNavigationBar = false {
+    var showsNavigationBar = false {
         
         didSet {
             for viewController in self.editorViewControllers {