+++ CotEditor/Sources/PatternSortViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2018 1024jp
+//  © 2018-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -44,6 +44,8 @@ final class PatternSortViewController: NSViewController, SortPatternViewControll
     
     /// keep tabViewController
     override func prepare(for segue: NSStoryboardSegue, sender: Any?) {
+        
+        super.prepare(for: segue, sender: sender)
         
         guard
             self.tabViewController == nil,