+++ Tests/StringExtensionsTests.swift
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2015-2018 1024jp
+//  © 2015-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -58,15 +58,6 @@ class StringExtensionsTests: XCTestCase {
         
         // single regional indicator
         XCTAssertEqual("🇦 ".count, 2)
-    }
-    
-    
-    func testWordsCount() {
-        
-        XCTAssertEqual("Clarus says moof!".numberOfWords, 3)
-        XCTAssertEqual("plain-text".numberOfWords, 2)
-        XCTAssertEqual("!".numberOfWords, 0)
-        XCTAssertEqual("".numberOfWords, 0)
     }